requirements.txt
openai
langfuse
python-dotenv
click
pydantic
asyncio
aiofiles